Clean up device channel serializer specific functionality into plugin.
