+++ kolibri/core/content/serializers.py
@@ -10,25 +10,7 @@ from kolibri.core.content.models import ChannelMetadata
 from kolibri.core.content.models import ContentNode
 from kolibri.core.content.models import File
 from kolibri.core.content.models import Language
-from kolibri.core.content.models import LocalFile
-from kolibri.core.content.utils.content_types_tools import (
-    renderable_contentnodes_without_topics_q_filter,
-)
 from kolibri.core.fields import create_timezonestamp
-
-
-def _files_for_nodes(nodes):
-    return LocalFile.objects.filter(files__contentnode__in=nodes)
-
-
-def _total_file_size(files_or_nodes):
-    if issubclass(files_or_nodes.model, LocalFile):
-        localfiles = files_or_nodes
-    elif issubclass(files_or_nodes.model, ContentNode):
-        localfiles = _files_for_nodes(files_or_nodes)
-    else:
-        raise TypeError("Expected queryset for LocalFile or ContentNode")
-    return localfiles.distinct().aggregate(Sum("file_size"))["file_size__sum"] or 0
 
 
 class DynamicFieldsModelSerializer(serializers.ModelSerializer):
@@ -53,52 +35,7 @@ class ChannelMetadataSerializer(serializers.ModelSerializer):
     lang_code = serializers.SerializerMethodField()
     lang_name = serializers.SerializerMethodField()
     available = serializers.SerializerMethodField()
-
-    def to_representation(self, instance):
-        # TODO: rtibbles - cleanup this for device specific serializer.
-        value = super(ChannelMetadataSerializer, self).to_representation(instance)
-
-        value.update({"num_coach_contents": instance.root.num_coach_contents})
-
-        # if the request includes a GET param 'include_fields', add the requested calculated fields
-        if "request" in self.context:
-
-            include_fields = (
-                self.context["request"].GET.get("include_fields", "").split(",")
-            )
-
-            if include_fields:
-
-                # build querysets for the full set of channel nodes, as well as those that are unrenderable
-                channel_nodes = ContentNode.objects.filter(channel_id=instance.id)
-                unrenderable_nodes = channel_nodes.exclude(
-                    renderable_contentnodes_without_topics_q_filter
-                )
-
-                if "total_resources" in include_fields:
-                    # count the total number of renderable non-topic resources in the channel
-                    # (note: it's faster to count them all and then subtract the unrenderables, of which there are fewer)
-                    value["total_resources"] = (
-                        channel_nodes.dedupe_by_content_id().count()
-                        - unrenderable_nodes.dedupe_by_content_id().count()
-                    )
-
-                if "total_file_size" in include_fields:
-                    # count the total file size of files associated with renderable content nodes
-                    # (note: it's faster to count them all and then subtract the unrenderables, of which there are fewer)
-                    value["total_file_size"] = _total_file_size(
-                        channel_nodes
-                    ) - _total_file_size(unrenderable_nodes)
-
-                if "on_device_resources" in include_fields:
-                    # read the precalculated total number of resources from the channel already available
-                    value["on_device_resources"] = instance.total_resource_count
-
-                if "on_device_file_size" in include_fields:
-                    # read the precalculated total size of available files associated with the channel
-                    value["on_device_file_size"] = instance.published_size
-
-        return value
+    num_coach_contents = serializers.IntegerField(source="root.num_coach_contents")
 
     def get_lang_code(self, instance):
         if instance.root.lang is None:
@@ -129,6 +66,7 @@ class ChannelMetadataSerializer(serializers.ModelSerializer):
             "thumbnail",
             "version",
             "available",
+            "num_coach_contents",
         )
