+++ kolibri/core/content/test/test_content_app.py
@@ -8,13 +8,11 @@ import mock
 import requests
 from django.core.cache import cache
 from django.core.urlresolvers import reverse
-from django.db.models import Q
 from django.utils import timezone
 from le_utils.constants import content_kinds
 from rest_framework import status
 from rest_framework.test import APITestCase
 
-import kolibri.core.content.serializers
 from kolibri.core.auth.constants import user_kinds
 from kolibri.core.auth.models import Facility
 from kolibri.core.auth.models import FacilityUser
@@ -820,27 +818,6 @@ class ContentNodeAPITestCase(APITestCase):
         )
         self.assertEqual(response.data["name"], "testing")
 
-    def test_channelmetadata_resource_info(self):
-        content.ChannelMetadata.objects.all().update(
-            total_resource_count=4, published_size=0
-        )
-        data = content.ChannelMetadata.objects.values()[0]
-        c1_id = content.ContentNode.objects.get(title="c1").id
-        content.ContentNode.objects.filter(pk=c1_id).update(available=False)
-        get_params = {
-            "include_fields": "total_resources,total_file_size,on_device_resources,on_device_file_size"
-        }
-        response = self.client.get(
-            reverse("kolibri:core:channel-detail", kwargs={"pk": data["id"]}),
-            get_params,
-        )
-        # N.B. Because of our not very good fixture data, most of our content nodes are by default not renderable
-        # Hence this will return 1 if everything is deduped properly.
-        self.assertEqual(response.data["total_resources"], 1)
-        self.assertEqual(response.data["total_file_size"], 0)
-        self.assertEqual(response.data["on_device_resources"], 4)
-        self.assertEqual(response.data["on_device_file_size"], 0)
-
     def test_channelmetadata_langfield(self):
         data = content.ChannelMetadata.objects.first()
         root_lang = content.Language.objects.get(pk=1)
@@ -889,64 +866,6 @@ class ContentNodeAPITestCase(APITestCase):
         content.ContentNode.objects.filter(title="root").update(available=False)
         response = self.client.get(reverse("kolibri:core:channel-list"))
         self.assertEqual(response.data[0]["available"], False)
-
-    def test_channelmetadata_include_fields_filter_has_total_resources(self):
-        # N.B. Because of our not very good fixture data, most of our content nodes are by default not renderable
-        # Hence this will return 1 if everything is deduped properly.
-        response = self.client.get(
-            reverse("kolibri:core:channel-list"), {"include_fields": "total_resources"}
-        )
-        self.assertEqual(response.data[0]["total_resources"], 1)
-
-    def test_channelmetadata_include_fields_filter_has_total_file_size(self):
-        content.LocalFile.objects.filter(
-            files__contentnode__channel_id=self.the_channel_id
-        ).update(file_size=1)
-        response = self.client.get(
-            reverse("kolibri:core:channel-list"), {"include_fields": "total_file_size"}
-        )
-        self.assertEqual(response.data[0]["total_file_size"], 5)
-
-    def test_channelmetadata_include_fields_filter_has_on_device_resources(self):
-        content.ChannelMetadata.objects.all().update(total_resource_count=5)
-        response = self.client.get(
-            reverse("kolibri:core:channel-list"),
-            {"include_fields": "on_device_resources"},
-        )
-        self.assertEqual(response.data[0]["on_device_resources"], 5)
-
-    def test_channelmetadata_include_fields_filter_has_on_device_file_size(self):
-        content.ChannelMetadata.objects.all().update(published_size=4)
-        response = self.client.get(
-            reverse("kolibri:core:channel-list"),
-            {"include_fields": "on_device_file_size"},
-        )
-        self.assertEqual(response.data[0]["on_device_file_size"], 4)
-
-    def test_channelmetadata_include_fields_filter_has_no_on_device_file_size(self):
-        content.ChannelMetadata.objects.all().update(published_size=0)
-        response = self.client.get(
-            reverse("kolibri:core:channel-list"),
-            {
-                "include_fields": "total_resources,total_file_size,on_device_resources,on_device_file_size"
-            },
-        )
-        self.assertEqual(response.data[0]["on_device_file_size"], 0)
-
-    @mock.patch.object(
-        kolibri.core.content.serializers,
-        "renderable_contentnodes_without_topics_q_filter",
-        Q(kind="dummy"),
-    )
-    def test_channelmetadata_include_fields_filter_has_no_renderable_on_device_file_size(
-        self,
-    ):
-        content.ChannelMetadata.objects.all().update(published_size=4)
-        response = self.client.get(
-            reverse("kolibri:core:channel-list"),
-            {"include_fields": "on_device_file_size"},
-        )
-        self.assertEqual(response.data[0]["on_device_file_size"], 4)
 
     def test_channelmetadata_has_exercises_filter(self):
         # Has nothing else for that matter...
